ml_service/integration_client.py
"""
🌐 INTEGRATION WEBSOCKET CLIENT
===============================

Connects Python ML services to the central Integration WebSocket Gateway
for seamless data flow and real-time communication.
"""

import asyncio
import json
import logging
from typing import Dict, Any, Optional, Callable
import socketio
from datetime import datetime

logger = logging.getLogger(__name__)


class IntegrationClient:
    """Client for connecting to the Integration WebSocket Gateway"""
    
    def __init__(self, 
                 service_name: str,
                 integration_url: str = "http://localhost:8888",
                 capabilities: list = None):
        """
        Initialize the integration client
        
        Args:
            service_name: Name of this service (e.g., 'ml_service', 'ai_coordination')
            integration_url: URL of the Integration WebSocket Gateway
            capabilities: List of capabilities this service provides
        """
        self.service_name = service_name
        self.integration_url = integration_url
        self.capabilities = capabilities or []
        
        # Create socket.io client
        self.sio = socketio.AsyncClient(
            reconnection=True,
            reconnection_delay=5,
            reconnection_delay_max=30,
            logger=logger
        )
        
        # Event handlers
        self._handlers: Dict[str, Callable] = {}
        
        # Setup default handlers
        self._setup_default_handlers()
        
    def _setup_default_handlers(self):
        """Setup default socket event handlers"""
        
        @self.sio.event
        async def connect():
            logger.info(f"✅ {self.service_name} connected to Integration Gateway")
            # Register this service
            await self.register_service()
            
        @self.sio.event
        async def connect_error(data):
            logger.error(f"❌ Connection error: {data}")
            
        @self.sio.event
        async def disconnect():
            logger.warning(f"🔌 {self.service_name} disconnected from Integration Gateway")
            
        @self.sio.event
        async def integration_status(data):
            logger.info(f"📊 Integration status: {data}")
            
        @self.sio.event
        async def game_data_update(data):
            """Handle game data updates"""
            if 'game_data_update' in self._handlers:
                await self._handlers['game_data_update'](data)
                
        @self.sio.event
        async def pattern_shared(data):
            """Handle shared patterns"""
            if 'pattern_shared' in self._handlers:
                await self._handlers['pattern_shared'](data)
                
        @self.sio.event
        async def model_updated(data):
            """Handle model updates"""
            if 'model_updated' in self._handlers:
                await self._handlers['model_updated'](data)
                
        @self.sio.event
        async def insight_available(data):
            """Handle new insights"""
            if 'insight_available' in self._handlers:
                await self._handlers['insight_available'](data)
                
        @self.sio.event
        async def service_status_update(data):
            """Handle service status updates"""
            if 'service_status_update' in self._handlers:
                await self._handlers['service_status_update'](data)
    
    async def connect(self):
        """Connect to the Integration Gateway"""
        try:
            await self.sio.connect(
                self.integration_url,
                namespaces=['/integration']
            )
            logger.info(f"🌐 Connecting {self.service_name} to Integration Gateway...")
        except Exception as e:
            logger.error(f"Failed to connect: {e}")
            raise
    
    async def disconnect(self):
        """Disconnect from the Integration Gateway"""
        await self.sio.disconnect()
    
    async def register_service(self):
        """Register this service with the Integration Gateway"""
        await self.emit('register_service', {
            'serviceName': self.service_name,
            'capabilities': self.capabilities,
            'timestamp': datetime.now().isoformat()
        })
    
    async def emit(self, event: str, data: Any):
        """Emit an event to the Integration Gateway"""
        try:
            await self.sio.emit(event, data, namespace='/integration')
        except Exception as e:
            logger.error(f"Failed to emit {event}: {e}")
    
    def on(self, event: str, handler: Callable):
        """Register an event handler"""
        self._handlers[event] = handler
    
    async def broadcast_game_data(self, game_data: Dict[str, Any]):
        """Broadcast game data to all services"""
        await self.emit('broadcast_game_data', game_data)
    
    async def share_pattern(self, pattern: Dict[str, Any]):
        """Share a detected pattern with all services"""
        await self.emit('share_pattern', {
            'pattern': pattern,
            'source': self.service_name,
            'timestamp': datetime.now().isoformat()
        })
    
    async def notify_model_update(self, model_type: str, version: str, 
                                  metadata: Dict[str, Any] = None):
        """Notify all services about a model update"""
        await self.emit('notify_model_update', {
            'modelType': model_type,
            'version': version,
            'metadata': metadata or {},
            'source': self.service_name,
            'timestamp': datetime.now().isoformat()
        })
    
    async def request_move_analysis(self, game_id: str, board: list, 
                                    move: Dict[str, Any]):
        """Request real-time move analysis from all services"""
        await self.emit('analyze_move_realtime', {
            'gameId': game_id,
            'board': board,
            'move': move,
            'source': self.service_name
        })
    
    async def submit_simulation_result(self, simulation_id: str, 
                                       result: Dict[str, Any]):
        """Submit results from AI vs AI simulation"""
        await self.emit('simulation_result', {
            'simulationId': simulation_id,
            'result': result,
            'source': self.service_name,
            'timestamp': datetime.now().isoformat()
        })
    
    async def propagate_insight(self, insight: Dict[str, Any]):
        """Propagate a strategic insight to all services"""
        await self.emit('propagate_insight', {
            'insight': insight,
            'source': self.service_name,
            'timestamp': datetime.now().isoformat()
        })
    
    async def request_metrics(self) -> Dict[str, Any]:
        """Request metrics from the Integration Gateway"""
        future = asyncio.Future()
        
        @self.sio.event
        async def metrics_response(data):
            future.set_result(data)
        
        await self.emit('request_metrics', {})
        
        try:
            return await asyncio.wait_for(future, timeout=5.0)
        except asyncio.TimeoutError:
            logger.error("Metrics request timed out")
            return {}


class MLServiceIntegration:
    """Integration wrapper for ML Service"""
    
    def __init__(self):
        self.client = IntegrationClient(
            service_name='ml_service',
            capabilities=[
                'model_training',
                'pattern_detection',
                'move_prediction',
                'performance_analysis'
            ]
        )
        
        # Setup ML-specific handlers
        self.client.on('game_data_update', self.handle_game_data)
        self.client.on('pattern_shared', self.handle_pattern)
        self.client.on('model_sync_request', self.handle_model_sync)
        
    async def handle_game_data(self, data: Dict[str, Any]):
        """Process game data for learning"""
        logger.info(f"📊 Processing game data: {data.get('gameId')}")
        # TODO: Implement game data processing
        
    async def handle_pattern(self, data: Dict[str, Any]):
        """Learn from shared patterns"""
        logger.info(f"🔍 Learning from pattern: {data.get('pattern', {}).get('type')}")
        # TODO: Implement pattern learning
        
    async def handle_model_sync(self, data: Dict[str, Any]):
        """Sync models across services"""
        logger.info(f"🔄 Syncing model: {data.get('modelType')}")
        # TODO: Implement model synchronization
        
    async def start(self):
        """Start the integration client"""
        await self.client.connect()
        logger.info("✅ ML Service Integration started")
        
    async def stop(self):
        """Stop the integration client"""
        await self.client.disconnect()
        logger.info("🛑 ML Service Integration stopped")


class AICoordinationIntegration:
    """Integration wrapper for AI Coordination Hub"""
    
    def __init__(self):
        self.client = IntegrationClient(
            service_name='ai_coordination',
            capabilities=[
                'strategic_analysis',
                'simulation_coordination',
                'insight_generation',
                'cross_ai_communication'
            ]
        )
        
        # Setup AI Coordination-specific handlers
        self.client.on('game_data_update', self.analyze_game_strategy)
        self.client.on('simulation_request', self.handle_simulation_request)
        
    async def analyze_game_strategy(self, data: Dict[str, Any]):
        """Analyze game for strategic insights"""
        logger.info(f"🎯 Analyzing strategy for game: {data.get('gameId')}")
        # TODO: Implement strategic analysis
        
    async def handle_simulation_request(self, data: Dict[str, Any]):
        """Handle request to run AI vs AI simulation"""
        logger.info(f"🎮 Running simulation: {data.get('simulationId')}")
        # TODO: Implement simulation coordination
        
    async def start(self):
        """Start the integration client"""
        await self.client.connect()
        logger.info("✅ AI Coordination Integration started")
        
    async def stop(self):
        """Stop the integration client"""
        await self.client.disconnect()
        logger.info("🛑 AI Coordination Integration stopped")


# Example usage
async def main():
    """Example of how to use the integration client"""
    
    # Create ML Service integration
    ml_integration = MLServiceIntegration()
    await ml_integration.start()
    
    # Simulate some events
    await ml_integration.client.notify_model_update(
        model_type='minimax',
        version='2.0.0',
        metadata={'accuracy': 0.92}
    )
    
    await ml_integration.client.share_pattern({
        'type': 'diagonal_threat',
        'positions': [[2, 3], [3, 4], [4, 5]],
        'confidence': 0.85
    })
    
    # Keep running
    try:
        await asyncio.Event().wait()
    except KeyboardInterrupt:
        await ml_integration.stop()


if __name__ == "__main__":
    logging.basicConfig(
        level=logging.INFO,
        format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
    )
    
    asyncio.run(main())