ml_service/requirements.txt
# Core ML and Web Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Data Science and Utilities
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Model Export and Optimization
onnx>=1.15.0
onnxruntime>=1.16.0

# Monitoring and Logging
prometheus-client>=0.19.0
structlog>=23.2.0
python-json-logger>=2.0.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.7.0

# Performance and Caching
redis>=5.0.0
aioredis>=2.0.0
aiocache>=0.12.0

# Security and Validation
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0

# HTTP and Async
httpx>=0.25.0
aiofiles>=23.2.0
python-multipart>=0.0.6
websockets>=12.0
python-socketio[asyncio]>=5.10.0
aiohttp>=3.9.0

# Environment and Configuration
python-dotenv>=1.0.0
